Blockly compiler.
Type-checking of blockly programs + error reporting.

All the blocks have been tagged with a type, including the "device" ones. The
types are: Boolean, Number, String, Image. The expected type is propagated down
when compiling the blocks. When there is a type mismatch, the block cannot be
compiled, so it is highlighted, an error is reported in the UI, and the user is
provided with a "force compile" / "force graduate" button along with an error
message (UI to be refined).

Normally, the very UI of Blockly prevents type mismatches (blocks themselves
have a basic type system). For instance, one cannot compare numbers and strings
(the block will just "pop out" when trying to do that). However, Blockly does
not perform type inference of variables. Therefore, our compiler performs a
quick & dirty type-checking pass that tries to infer the type of variables,
using a really dumb fixpoint computation (given the typical size of Blockly
programs, it probably does not warrant writing a Hindley-Milner inference system
in Typescript).

All blockly variables are now compiled as globals, except for loop indices,
which are compiled as usual. Loop indices that are assigned to generate a global
along with a while loop, meaning that the translation remains correct.